requirements.txt
PyQt5>=5.15.7
torch>=2.0.0
transformers>=4.30.0
sentencepiece>=0.1.99
openai-whisper>=20230101
pyttsx3>=2.90
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.23.0
scipy>=1.9.0
librosa>=0.9.2
python-dotenv>=1.0.0
requests>=2.31.0
PyYAML>=6.0
Pillow>=9.0.0
tqdm>=4.65.0
